Run STFTs in float32 so bf16 training works on CUDA
cuFFT has no half/bfloat16 kernels, so the multi-resolution STFT discriminator
and the shared spectrogram front-end crashed under bf16-mixed precision. Both
now upcast before the transform; autocast casts back for the convolutions.
Adds regression tests for reduced-precision input.

+++ src/auris_singer/modules/discriminator.py
@@ -162,7 +162,9 @@ class STFTDiscriminator(nn.Module):
         self.projection = _ProjectionConditioning(n_speakers, channels)
 
     def _stft(self, x: torch.Tensor) -> torch.Tensor:
-        x = x.squeeze(1)
+        # cuFFT has no half/bfloat16 kernels, so the transform always runs in
+        # float32; under autocast the following convolutions cast it back.
+        x = x.squeeze(1).float()
         pad = (self.n_fft - self.hop_length) // 2
         x = F.pad(x.unsqueeze(1), (pad, pad), mode="reflect").squeeze(1)
         spec = torch.stft(
@@ -170,7 +172,7 @@ class STFTDiscriminator(nn.Module):
             n_fft=self.n_fft,
             hop_length=self.hop_length,
             win_length=self.win_length,
-            window=self.window.to(x.dtype),
+            window=self.window.float(),
             center=False,
             return_complex=True,
         )
